refactor: LottoCalculator 클래스 분리하고 적용
