+++ src/model/index.js
@@ -1,5 +1,6 @@
 export { default as Lotto } from "./Lotto.js";
 export { default as LottoMachine } from "./LottoMachine.js";
 export { default as WinningLotto } from "./WinningLotto.js";
-export { default as LottoCalculator } from "./LottoCalculator.js";
 export { default as LottoNumber } from "./LottoNumber.js";
+export { default as LottoMatcher } from "./LottoMatcher.js";
+export { default as LottoResult } from "./LottoResult.js";
